extract shared conflicts-based work into own module
