+++ app/src/ui/lib/conflicts/index.ts
@@ -1,0 +1,2 @@
+export * from './render-functions'
+export * from './unmerged-file'
